confluence-rag/requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
atlassian-python-api
python-dotenv
tiktoken
beautifulsoup4
lxml
